more changes which result in imports working. by haivng them all in the same file. pooo
